+++ dpws/src/test/java/org/somda/sdc/dpws/CommunicationLogImplTest.java
@@ -22,11 +22,11 @@ public class CommunicationLogImplTest extends DpwsTest {
 
         byte[] content = UUID.randomUUID().toString().getBytes();
 
-        try (ByteArrayOutputStream mockOutputStream = new ByteArrayOutputStream();
+        try (ByteArrayOutputStream mockOutputStream = spy(new ByteArrayOutputStream());
              ByteArrayInputStream inputTestInputStream = new ByteArrayInputStream(content);
              ByteArrayOutputStream outputTestOutputStream = new ByteArrayOutputStream();) {
 
-            when(communicationLogSinkImplMock.getTargetStream(eq(CommunicationLog.TransportType.HTTP), any(), any()))
+            when(communicationLogSinkImplMock.createTargetStream(eq(CommunicationLog.TransportType.HTTP), any(), any()))
                     .thenReturn(mockOutputStream);
 
             CommunicationLogImpl communicationLogImpl = new CommunicationLogImpl(communicationLogSinkImplMock);
@@ -47,6 +47,7 @@ public class CommunicationLogImplTest extends DpwsTest {
 
             assertArrayEquals(resultingInputStream.readAllBytes(), content);
             assertArrayEquals(mockOutputStream.toByteArray(), content);
+            verify(mockOutputStream, times(1)).close();
 
             mockOutputStream.reset();
 
@@ -86,7 +87,7 @@ public class CommunicationLogImplTest extends DpwsTest {
             for (CommunicationLog.Direction dir : CommunicationLog.Direction.values()) {
                 reset(communicationLogSinkImplMock);
 
-                when(communicationLogSinkImplMock.getTargetStream(any(CommunicationLog.TransportType.class), any(), any()))
+                when(communicationLogSinkImplMock.createTargetStream(any(CommunicationLog.TransportType.class), any(), any()))
                         .thenReturn(OutputStream.nullOutputStream());
 
                 communicationLogImpl.logMessage(dir, CommunicationLog.TransportType.HTTP, requestCommContext,
@@ -94,7 +95,7 @@ public class CommunicationLogImplTest extends DpwsTest {
                 communicationLogImpl.logMessage(dir, CommunicationLog.TransportType.HTTP, requestCommContext,
                         OutputStream.nullOutputStream());
                 verify(communicationLogSinkImplMock, times(2)).
-                        getTargetStream(eq(CommunicationLog.TransportType.HTTP), any(), any());
+                        createTargetStream(eq(CommunicationLog.TransportType.HTTP), any(), any());
             }
         }
     }